% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_write}
\alias{cache_write}
\title{Write a panel to the cache}
\usage{
cache_write(panel, source, vintage)
}
\arguments{
\item{panel}{A tibble. Must have columns
\verb{(series_id, source, date, value, vintage)}.}

\item{source}{A source name (see \code{\link[=cache_file_path]{cache_file_path()}}).}

\item{vintage}{A \code{Date}.}
}
\value{
The input \code{panel} invisibly.
}
\description{
Write a panel to the cache
}
